+++ projects/android/koin-android/src/main/java/org/koin/androidx/viewmodel/dsl/ViewModelOf.kt
@@ -42,6 +42,7 @@ import org.koin.core.module.dsl.onOptions
  *
  * @see new
  */
+@Deprecated("Moved ViewModel DSL package. Remove old imports and use org.koin.core.module.dsl.*")
 inline fun <reified R : ViewModel> Module.viewModelOf(
     crossinline constructor: () -> R,
     noinline options: DefinitionOptions<R>? = null,
@@ -50,6 +51,7 @@ inline fun <reified R : ViewModel> Module.viewModelOf(
 /**
  * @see viewModelOf
  */
+@Deprecated("Moved ViewModel DSL package. Remove old imports and use org.koin.core.module.dsl.*")
 inline fun <reified R : ViewModel, reified T1> Module.viewModelOf(
     crossinline constructor: (T1) -> R,
     noinline options: DefinitionOptions<R>? = null,
@@ -58,6 +60,7 @@ inline fun <reified R : ViewModel, reified T1> Module.viewModelOf(
 /**
  * @see viewModelOf
  */
+@Deprecated("Moved ViewModel DSL package. Remove old imports and use org.koin.core.module.dsl.*")
 inline fun <reified R : ViewModel, reified T1, reified T2> Module.viewModelOf(
     crossinline constructor: (T1, T2) -> R,
     noinline options: DefinitionOptions<R>? = null,
@@ -66,6 +69,7 @@ inline fun <reified R : ViewModel, reified T1, reified T2> Module.viewModelOf(
 /**
  * @see viewModelOf
  */
+@Deprecated("Moved ViewModel DSL package. Remove old imports and use org.koin.core.module.dsl.*")
 inline fun <reified R : ViewModel, reified T1, reified T2, reified T3> Module.viewModelOf(
     crossinline constructor: (T1, T2, T3) -> R,
     noinline options: DefinitionOptions<R>? = null,
@@ -74,6 +78,7 @@ inline fun <reified R : ViewModel, reified T1, reified T2, reified T3> Module.vi
 /**
  * @see viewModelOf
  */
+@Deprecated("Moved ViewModel DSL package. Remove old imports and use org.koin.core.module.dsl.*")
 inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T4> Module.viewModelOf(
     crossinline constructor: (T1, T2, T3, T4) -> R,
     noinline options: DefinitionOptions<R>? = null,
@@ -82,6 +87,7 @@ inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T
 /**
  * @see viewModelOf
  */
+@Deprecated("Moved ViewModel DSL package. Remove old imports and use org.koin.core.module.dsl.*")
 inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T4, reified T5> Module.viewModelOf(
     crossinline constructor: (T1, T2, T3, T4, T5) -> R,
     noinline options: DefinitionOptions<R>? = null,
@@ -90,6 +96,7 @@ inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T
 /**
  * @see viewModelOf
  */
+@Deprecated("Moved ViewModel DSL package. Remove old imports and use org.koin.core.module.dsl.*")
 inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T4, reified T5, reified T6> Module.viewModelOf(
     crossinline constructor: (T1, T2, T3, T4, T5, T6) -> R,
     noinline options: DefinitionOptions<R>? = null,
@@ -98,6 +105,7 @@ inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T
 /**
  * @see viewModelOf
  */
+@Deprecated("Moved ViewModel DSL package. Remove old imports and use org.koin.core.module.dsl.*")
 inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T4, reified T5, reified T6, reified T7> Module.viewModelOf(
     crossinline constructor: (T1, T2, T3, T4, T5, T6, T7) -> R,
     noinline options: DefinitionOptions<R>? = null,
@@ -106,6 +114,7 @@ inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T
 /**
  * @see viewModelOf
  */
+@Deprecated("Moved ViewModel DSL package. Remove old imports and use org.koin.core.module.dsl.*")
 inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T4, reified T5, reified T6, reified T7, reified T8> Module.viewModelOf(
     crossinline constructor: (T1, T2, T3, T4, T5, T6, T7, T8) -> R,
     noinline options: DefinitionOptions<R>? = null,
@@ -114,6 +123,7 @@ inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T
 /**
  * @see viewModelOf
  */
+@Deprecated("Moved ViewModel DSL package. Remove old imports and use org.koin.core.module.dsl.*")
 inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T4, reified T5, reified T6, reified T7, reified T8, reified T9> Module.viewModelOf(
     crossinline constructor: (T1, T2, T3, T4, T5, T6, T7, T8, T9) -> R,
     noinline options: DefinitionOptions<R>? = null,
@@ -122,6 +132,7 @@ inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T
 /**
  * @see viewModelOf
  */
+@Deprecated("Moved ViewModel DSL package. Remove old imports and use org.koin.core.module.dsl.*")
 inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T4, reified T5, reified T6, reified T7, reified T8, reified T9, reified T10> Module.viewModelOf(
     crossinline constructor: (T1, T2, T3, T4, T5, T6, T7, T8, T9, T10) -> R,
     noinline options: DefinitionOptions<R>? = null,
@@ -130,6 +141,7 @@ inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T
 /**
  * @see viewModelOf
  */
+@Deprecated("Moved ViewModel DSL package. Remove old imports and use org.koin.core.module.dsl.*")
 inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T4, reified T5, reified T6, reified T7, reified T8, reified T9, reified T10, reified T11> Module.viewModelOf(
     crossinline constructor: (T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11) -> R,
     noinline options: DefinitionOptions<R>? = null,
@@ -138,6 +150,7 @@ inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T
 /**
  * @see viewModelOf
  */
+@Deprecated("Moved ViewModel DSL package. Remove old imports and use org.koin.core.module.dsl.*")
 inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T4, reified T5, reified T6, reified T7, reified T8, reified T9, reified T10, reified T11, reified T12> Module.viewModelOf(
     crossinline constructor: (T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12) -> R,
     noinline options: DefinitionOptions<R>? = null,
@@ -146,6 +159,7 @@ inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T
 /**
  * @see viewModelOf
  */
+@Deprecated("Moved ViewModel DSL package. Remove old imports and use org.koin.core.module.dsl.*")
 inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T4, reified T5, reified T6, reified T7, reified T8, reified T9, reified T10, reified T11, reified T12, reified T13> Module.viewModelOf(
     crossinline constructor: (T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13) -> R,
     noinline options: DefinitionOptions<R>? = null,
@@ -154,6 +168,7 @@ inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T
 /**
  * @see viewModelOf
  */
+@Deprecated("Moved ViewModel DSL package. Remove old imports and use org.koin.core.module.dsl.*")
 inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T4, reified T5, reified T6, reified T7, reified T8, reified T9, reified T10, reified T11, reified T12, reified T13, reified T14> Module.viewModelOf(
     crossinline constructor: (T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14) -> R,
     noinline options: DefinitionOptions<R>? = null,
@@ -162,6 +177,7 @@ inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T
 /**
  * @see viewModelOf
  */
+@Deprecated("Moved ViewModel DSL package. Remove old imports and use org.koin.core.module.dsl.*")
 inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T4, reified T5, reified T6, reified T7, reified T8, reified T9, reified T10, reified T11, reified T12, reified T13, reified T14, reified T15> Module.viewModelOf(
     crossinline constructor: (T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14, T15) -> R,
     noinline options: DefinitionOptions<R>? = null,
@@ -170,6 +186,7 @@ inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T
 /**
  * @see viewModelOf
  */
+@Deprecated("Moved ViewModel DSL package. Remove old imports and use org.koin.core.module.dsl.*")
 inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T4, reified T5, reified T6, reified T7, reified T8, reified T9, reified T10, reified T11, reified T12, reified T13, reified T14, reified T15, reified T16> Module.viewModelOf(
     crossinline constructor: (T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14, T15, T16) -> R,
     noinline options: DefinitionOptions<R>? = null,
@@ -178,6 +195,7 @@ inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T
 /**
  * @see viewModelOf
  */
+@Deprecated("Moved ViewModel DSL package. Remove old imports and use org.koin.core.module.dsl.*")
 inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T4, reified T5, reified T6, reified T7, reified T8, reified T9, reified T10, reified T11, reified T12, reified T13, reified T14, reified T15, reified T16, reified T17> Module.viewModelOf(
     crossinline constructor: (T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14, T15, T16, T17) -> R,
     noinline options: DefinitionOptions<R>? = null,
@@ -186,6 +204,7 @@ inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T
 /**
  * @see viewModelOf
  */
+@Deprecated("Moved ViewModel DSL package. Remove old imports and use org.koin.core.module.dsl.*")
 inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T4, reified T5, reified T6, reified T7, reified T8, reified T9, reified T10, reified T11, reified T12, reified T13, reified T14, reified T15, reified T16, reified T17, reified T18> Module.viewModelOf(
     crossinline constructor: (T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14, T15, T16, T17, T18) -> R,
     noinline options: DefinitionOptions<R>? = null,
@@ -194,6 +213,7 @@ inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T
 /**
  * @see viewModelOf
  */
+@Deprecated("Moved ViewModel DSL package. Remove old imports and use org.koin.core.module.dsl.*")
 inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T4, reified T5, reified T6, reified T7, reified T8, reified T9, reified T10, reified T11, reified T12, reified T13, reified T14, reified T15, reified T16, reified T17, reified T18, reified T19> Module.viewModelOf(
     crossinline constructor: (T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14, T15, T16, T17, T18, T19) -> R,
     noinline options: DefinitionOptions<R>? = null,
@@ -202,6 +222,7 @@ inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T
 /**
  * @see viewModelOf
  */
+@Deprecated("Moved ViewModel DSL package. Remove old imports and use org.koin.core.module.dsl.*")
 inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T4, reified T5, reified T6, reified T7, reified T8, reified T9, reified T10, reified T11, reified T12, reified T13, reified T14, reified T15, reified T16, reified T17, reified T18, reified T19, reified T20> Module.viewModelOf(
     crossinline constructor: (T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14, T15, T16, T17, T18, T19, T20) -> R,
     noinline options: DefinitionOptions<R>? = null,
@@ -210,6 +231,7 @@ inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T
 /**
  * @see viewModelOf
  */
+@Deprecated("Moved ViewModel DSL package. Remove old imports and use org.koin.core.module.dsl.*")
 inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T4, reified T5, reified T6, reified T7, reified T8, reified T9, reified T10, reified T11, reified T12, reified T13, reified T14, reified T15, reified T16, reified T17, reified T18, reified T19, reified T20, reified T21> Module.viewModelOf(
     crossinline constructor: (T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14, T15, T16, T17, T18, T19, T20, T21) -> R,
     noinline options: DefinitionOptions<R>? = null,
@@ -218,6 +240,7 @@ inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T
 /**
  * @see viewModelOf
  */
+@Deprecated("Moved ViewModel DSL package. Remove old imports and use org.koin.core.module.dsl.*")
 inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T4, reified T5, reified T6, reified T7, reified T8, reified T9, reified T10, reified T11, reified T12, reified T13, reified T14, reified T15, reified T16, reified T17, reified T18, reified T19, reified T20, reified T21, reified T22> Module.viewModelOf(
     crossinline constructor: (T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14, T15, T16, T17, T18, T19, T20, T21, T22) -> R,
     noinline options: DefinitionOptions<R>? = null,
