+++ projects/compose/koin-compose-viewmodel/src/commonMain/kotlin/org/koin/compose/viewmodel/dsl/ScopeSetExt.kt
@@ -32,7 +32,7 @@ import org.koin.dsl.ScopeDSL
  * @param qualifier - definition qualifier
  * @param definition - allow definition override
  */
-@Deprecated("use org.koin.viewmodel.dsl", ReplaceWith(expression = "viewModel", imports = ["org.koin.compose.dsl"]))
+@Deprecated("Moved ViewModel DSL package. Remove old imports and use org.koin.core.module.dsl.*")
 inline fun <reified T : ViewModel> ScopeDSL.viewModel(
     qualifier: Qualifier? = null,
     noinline definition: Definition<T>
