+++ projects/compose/koin-compose-viewmodel/src/commonMain/kotlin/org/koin/compose/viewmodel/dsl/ScopeViewModelOf.kt
@@ -43,7 +43,7 @@ import org.koin.dsl.ScopeDSL
  *
  * @see new - function operator
  */
-@Deprecated("use org.koin.viewmodel.dsl", ReplaceWith(expression = "viewModel", imports = ["org.koin.compose.dsl"]))
+@Deprecated("Moved ViewModel DSL package. Remove old imports and use org.koin.core.module.dsl.*")
 inline fun <reified R : ViewModel> ScopeDSL.viewModelOf(
     crossinline constructor: () -> R,
     noinline options: DefinitionOptions<R>? = null,
@@ -52,7 +52,7 @@ inline fun <reified R : ViewModel> ScopeDSL.viewModelOf(
 /**
  * @see viewModelOf
  */
-@Deprecated("use org.koin.viewmodel.dsl", ReplaceWith(expression = "viewModel", imports = ["org.koin.compose.dsl"]))
+@Deprecated("Moved ViewModel DSL package. Remove old imports and use org.koin.core.module.dsl.*")
 inline fun <reified R : ViewModel, reified T1> ScopeDSL.viewModelOf(
     crossinline constructor: (T1) -> R,
     noinline options: DefinitionOptions<R>? = null,
@@ -61,7 +61,7 @@ inline fun <reified R : ViewModel, reified T1> ScopeDSL.viewModelOf(
 /**
  * @see viewModelOf
  */
-@Deprecated("use org.koin.viewmodel.dsl", ReplaceWith(expression = "viewModel", imports = ["org.koin.compose.dsl"]))
+@Deprecated("Moved ViewModel DSL package. Remove old imports and use org.koin.core.module.dsl.*")
 inline fun <reified R : ViewModel, reified T1, reified T2> ScopeDSL.viewModelOf(
     crossinline constructor: (T1, T2) -> R,
     noinline options: DefinitionOptions<R>? = null,
@@ -70,7 +70,7 @@ inline fun <reified R : ViewModel, reified T1, reified T2> ScopeDSL.viewModelOf(
 /**
  * @see viewModelOf
  */
-@Deprecated("use org.koin.viewmodel.dsl", ReplaceWith(expression = "viewModel", imports = ["org.koin.compose.dsl"]))
+@Deprecated("Moved ViewModel DSL package. Remove old imports and use org.koin.core.module.dsl.*")
 inline fun <reified R : ViewModel, reified T1, reified T2, reified T3> ScopeDSL.viewModelOf(
     crossinline constructor: (T1, T2, T3) -> R,
     noinline options: DefinitionOptions<R>? = null,
@@ -79,7 +79,7 @@ inline fun <reified R : ViewModel, reified T1, reified T2, reified T3> ScopeDSL.
 /**
  * @see viewModelOf
  */
-@Deprecated("use org.koin.viewmodel.dsl", ReplaceWith(expression = "viewModel", imports = ["org.koin.compose.dsl"]))
+@Deprecated("Moved ViewModel DSL package. Remove old imports and use org.koin.core.module.dsl.*")
 inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T4> ScopeDSL.viewModelOf(
     crossinline constructor: (T1, T2, T3, T4) -> R,
     noinline options: DefinitionOptions<R>? = null,
@@ -88,7 +88,7 @@ inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T
 /**
  * @see viewModelOf
  */
-@Deprecated("use org.koin.viewmodel.dsl", ReplaceWith(expression = "viewModel", imports = ["org.koin.compose.dsl"]))
+@Deprecated("Moved ViewModel DSL package. Remove old imports and use org.koin.core.module.dsl.*")
 inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T4, reified T5> ScopeDSL.viewModelOf(
     crossinline constructor: (T1, T2, T3, T4, T5) -> R,
     noinline options: DefinitionOptions<R>? = null,
@@ -97,7 +97,7 @@ inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T
 /**
  * @see viewModelOf
  */
-@Deprecated("use org.koin.viewmodel.dsl", ReplaceWith(expression = "viewModel", imports = ["org.koin.compose.dsl"]))
+@Deprecated("Moved ViewModel DSL package. Remove old imports and use org.koin.core.module.dsl.*")
 inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T4, reified T5, reified T6> ScopeDSL.viewModelOf(
     crossinline constructor: (T1, T2, T3, T4, T5, T6) -> R,
     noinline options: DefinitionOptions<R>? = null,
@@ -106,7 +106,7 @@ inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T
 /**
  * @see viewModelOf
  */
-@Deprecated("use org.koin.viewmodel.dsl", ReplaceWith(expression = "viewModel", imports = ["org.koin.compose.dsl"]))
+@Deprecated("Moved ViewModel DSL package. Remove old imports and use org.koin.core.module.dsl.*")
 inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T4, reified T5, reified T6, reified T7> ScopeDSL.viewModelOf(
     crossinline constructor: (T1, T2, T3, T4, T5, T6, T7) -> R,
     noinline options: DefinitionOptions<R>? = null,
@@ -115,7 +115,7 @@ inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T
 /**
  * @see viewModelOf
  */
-@Deprecated("use org.koin.viewmodel.dsl", ReplaceWith(expression = "viewModel", imports = ["org.koin.compose.dsl"]))
+@Deprecated("Moved ViewModel DSL package. Remove old imports and use org.koin.core.module.dsl.*")
 inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T4, reified T5, reified T6, reified T7, reified T8> ScopeDSL.viewModelOf(
     crossinline constructor: (T1, T2, T3, T4, T5, T6, T7, T8) -> R,
     noinline options: DefinitionOptions<R>? = null,
@@ -124,7 +124,7 @@ inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T
 /**
  * @see viewModelOf
  */
-@Deprecated("use org.koin.viewmodel.dsl", ReplaceWith(expression = "viewModel", imports = ["org.koin.compose.dsl"]))
+@Deprecated("Moved ViewModel DSL package. Remove old imports and use org.koin.core.module.dsl.*")
 inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T4, reified T5, reified T6, reified T7, reified T8, reified T9> ScopeDSL.viewModelOf(
     crossinline constructor: (T1, T2, T3, T4, T5, T6, T7, T8, T9) -> R,
     noinline options: DefinitionOptions<R>? = null,
@@ -133,7 +133,7 @@ inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T
 /**
  * @see viewModelOf
  */
-@Deprecated("use org.koin.viewmodel.dsl", ReplaceWith(expression = "viewModel", imports = ["org.koin.compose.dsl"]))
+@Deprecated("Moved ViewModel DSL package. Remove old imports and use org.koin.core.module.dsl.*")
 inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T4, reified T5, reified T6, reified T7, reified T8, reified T9, reified T10> ScopeDSL.viewModelOf(
     crossinline constructor: (T1, T2, T3, T4, T5, T6, T7, T8, T9, T10) -> R,
     noinline options: DefinitionOptions<R>? = null,
@@ -142,7 +142,7 @@ inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T
 /**
  * @see viewModelOf
  */
-@Deprecated("use org.koin.viewmodel.dsl", ReplaceWith(expression = "viewModel", imports = ["org.koin.compose.dsl"]))
+@Deprecated("Moved ViewModel DSL package. Remove old imports and use org.koin.core.module.dsl.*")
 inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T4, reified T5, reified T6, reified T7, reified T8, reified T9, reified T10, reified T11> ScopeDSL.viewModelOf(
     crossinline constructor: (T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11) -> R,
     noinline options: DefinitionOptions<R>? = null,
@@ -151,7 +151,7 @@ inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T
 /**
  * @see viewModelOf
  */
-@Deprecated("use org.koin.viewmodel.dsl", ReplaceWith(expression = "viewModel", imports = ["org.koin.compose.dsl"]))
+@Deprecated("Moved ViewModel DSL package. Remove old imports and use org.koin.core.module.dsl.*")
 inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T4, reified T5, reified T6, reified T7, reified T8, reified T9, reified T10, reified T11, reified T12> ScopeDSL.viewModelOf(
     crossinline constructor: (T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12) -> R,
     noinline options: DefinitionOptions<R>? = null,
@@ -160,7 +160,7 @@ inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T
 /**
  * @see viewModelOf
  */
-@Deprecated("use org.koin.viewmodel.dsl", ReplaceWith(expression = "viewModel", imports = ["org.koin.compose.dsl"]))
+@Deprecated("Moved ViewModel DSL package. Remove old imports and use org.koin.core.module.dsl.*")
 inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T4, reified T5, reified T6, reified T7, reified T8, reified T9, reified T10, reified T11, reified T12, reified T13> ScopeDSL.viewModelOf(
     crossinline constructor: (T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13) -> R,
     noinline options: DefinitionOptions<R>? = null,
@@ -169,7 +169,7 @@ inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T
 /**
  * @see viewModelOf
  */
-@Deprecated("use org.koin.viewmodel.dsl", ReplaceWith(expression = "viewModel", imports = ["org.koin.compose.dsl"]))
+@Deprecated("Moved ViewModel DSL package. Remove old imports and use org.koin.core.module.dsl.*")
 inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T4, reified T5, reified T6, reified T7, reified T8, reified T9, reified T10, reified T11, reified T12, reified T13, reified T14> ScopeDSL.viewModelOf(
     crossinline constructor: (T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14) -> R,
     noinline options: DefinitionOptions<R>? = null,
@@ -178,7 +178,7 @@ inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T
 /**
  * @see viewModelOf
  */
-@Deprecated("use org.koin.viewmodel.dsl", ReplaceWith(expression = "viewModel", imports = ["org.koin.compose.dsl"]))
+@Deprecated("Moved ViewModel DSL package. Remove old imports and use org.koin.core.module.dsl.*")
 inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T4, reified T5, reified T6, reified T7, reified T8, reified T9, reified T10, reified T11, reified T12, reified T13, reified T14, reified T15> ScopeDSL.viewModelOf(
     crossinline constructor: (T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14, T15) -> R,
     noinline options: DefinitionOptions<R>? = null,
@@ -187,7 +187,7 @@ inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T
 /**
  * @see viewModelOf
  */
-@Deprecated("use org.koin.viewmodel.dsl", ReplaceWith(expression = "viewModel", imports = ["org.koin.compose.dsl"]))
+@Deprecated("Moved ViewModel DSL package. Remove old imports and use org.koin.core.module.dsl.*")
 inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T4, reified T5, reified T6, reified T7, reified T8, reified T9, reified T10, reified T11, reified T12, reified T13, reified T14, reified T15, reified T16> ScopeDSL.viewModelOf(
     crossinline constructor: (T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14, T15, T16) -> R,
     noinline options: DefinitionOptions<R>? = null,
@@ -196,7 +196,7 @@ inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T
 /**
  * @see viewModelOf
  */
-@Deprecated("use org.koin.viewmodel.dsl", ReplaceWith(expression = "viewModel", imports = ["org.koin.compose.dsl"]))
+@Deprecated("Moved ViewModel DSL package. Remove old imports and use org.koin.core.module.dsl.*")
 inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T4, reified T5, reified T6, reified T7, reified T8, reified T9, reified T10, reified T11, reified T12, reified T13, reified T14, reified T15, reified T16, reified T17> ScopeDSL.viewModelOf(
     crossinline constructor: (T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14, T15, T16, T17) -> R,
     noinline options: DefinitionOptions<R>? = null,
@@ -205,7 +205,7 @@ inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T
 /**
  * @see viewModelOf
  */
-@Deprecated("use org.koin.viewmodel.dsl", ReplaceWith(expression = "viewModel", imports = ["org.koin.compose.dsl"]))
+@Deprecated("Moved ViewModel DSL package. Remove old imports and use org.koin.core.module.dsl.*")
 inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T4, reified T5, reified T6, reified T7, reified T8, reified T9, reified T10, reified T11, reified T12, reified T13, reified T14, reified T15, reified T16, reified T17, reified T18> ScopeDSL.viewModelOf(
     crossinline constructor: (T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14, T15, T16, T17, T18) -> R,
     noinline options: DefinitionOptions<R>? = null,
@@ -214,7 +214,7 @@ inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T
 /**
  * @see viewModelOf
  */
-@Deprecated("use org.koin.viewmodel.dsl", ReplaceWith(expression = "viewModel", imports = ["org.koin.compose.dsl"]))
+@Deprecated("Moved ViewModel DSL package. Remove old imports and use org.koin.core.module.dsl.*")
 inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T4, reified T5, reified T6, reified T7, reified T8, reified T9, reified T10, reified T11, reified T12, reified T13, reified T14, reified T15, reified T16, reified T17, reified T18, reified T19> ScopeDSL.viewModelOf(
     crossinline constructor: (T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14, T15, T16, T17, T18, T19) -> R,
     noinline options: DefinitionOptions<R>? = null,
@@ -223,7 +223,7 @@ inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T
 /**
  * @see viewModelOf
  */
-@Deprecated("use org.koin.viewmodel.dsl", ReplaceWith(expression = "viewModel", imports = ["org.koin.compose.dsl"]))
+@Deprecated("Moved ViewModel DSL package. Remove old imports and use org.koin.core.module.dsl.*")
 inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T4, reified T5, reified T6, reified T7, reified T8, reified T9, reified T10, reified T11, reified T12, reified T13, reified T14, reified T15, reified T16, reified T17, reified T18, reified T19, reified T20> ScopeDSL.viewModelOf(
     crossinline constructor: (T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14, T15, T16, T17, T18, T19, T20) -> R,
     noinline options: DefinitionOptions<R>? = null,
@@ -232,7 +232,7 @@ inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T
 /**
  * @see viewModelOf
  */
-@Deprecated("use org.koin.viewmodel.dsl", ReplaceWith(expression = "viewModel", imports = ["org.koin.compose.dsl"]))
+@Deprecated("Moved ViewModel DSL package. Remove old imports and use org.koin.core.module.dsl.*")
 inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T4, reified T5, reified T6, reified T7, reified T8, reified T9, reified T10, reified T11, reified T12, reified T13, reified T14, reified T15, reified T16, reified T17, reified T18, reified T19, reified T20, reified T21> ScopeDSL.viewModelOf(
     crossinline constructor: (T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14, T15, T16, T17, T18, T19, T20, T21) -> R,
     noinline options: DefinitionOptions<R>? = null,
@@ -241,7 +241,7 @@ inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T
 /**
  * @see viewModelOf
  */
-@Deprecated("use org.koin.viewmodel.dsl", ReplaceWith(expression = "viewModel", imports = ["org.koin.compose.dsl"]))
+@Deprecated("Moved ViewModel DSL package. Remove old imports and use org.koin.core.module.dsl.*")
 inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T4, reified T5, reified T6, reified T7, reified T8, reified T9, reified T10, reified T11, reified T12, reified T13, reified T14, reified T15, reified T16, reified T17, reified T18, reified T19, reified T20, reified T21, reified T22> ScopeDSL.viewModelOf(
     crossinline constructor: (T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14, T15, T16, T17, T18, T19, T20, T21, T22) -> R,
     noinline options: DefinitionOptions<R>? = null,
