+++ projects/core/koin-core-viewmodel/src/commonMain/kotlin/org/koin/core/module/dsl/ScopeViewModelOf.kt
@@ -14,17 +14,14 @@
  * limitations under the License.
  */
 
-package org.koin.viewmodel.dsl
+package org.koin.core.module.dsl
 
 import androidx.lifecycle.ViewModel
 import org.koin.core.definition.KoinDefinition
-import org.koin.core.module.Module
-import org.koin.core.module.dsl.DefinitionOptions
-import org.koin.core.module.dsl.new
-import org.koin.core.module.dsl.onOptions
+import org.koin.dsl.ScopeDSL
 
 /**
- * Declare a [Module.viewModel] definition by resolving a constructor reference for the dependency.
+ * Declare a [ScopeDSL.viewModel] definition by resolving a constructor reference for the dependency.
  * The resolution is done at compile time by leveraging inline functions, no reflection is required.
  *
  * Example:
@@ -38,9 +35,10 @@ import org.koin.core.module.dsl.onOptions
  *
  * @author Arnaud Giuliani
  *
- * @see new
+ * @see new - function operator
  */
-inline fun <reified R : ViewModel> Module.viewModelOf(
+
+inline fun <reified R : ViewModel> ScopeDSL.viewModelOf(
     crossinline constructor: () -> R,
     noinline options: DefinitionOptions<R>? = null,
 ): KoinDefinition<R> = viewModel { new(constructor) }.onOptions(options)
@@ -48,7 +46,7 @@ inline fun <reified R : ViewModel> Module.viewModelOf(
 /**
  * @see viewModelOf
  */
-inline fun <reified R : ViewModel, reified T1> Module.viewModelOf(
+inline fun <reified R : ViewModel, reified T1> ScopeDSL.viewModelOf(
     crossinline constructor: (T1) -> R,
     noinline options: DefinitionOptions<R>? = null,
 ): KoinDefinition<R> = viewModel { new(constructor) }.onOptions(options)
@@ -56,7 +54,7 @@ inline fun <reified R : ViewModel, reified T1> Module.viewModelOf(
 /**
  * @see viewModelOf
  */
-inline fun <reified R : ViewModel, reified T1, reified T2> Module.viewModelOf(
+inline fun <reified R : ViewModel, reified T1, reified T2> ScopeDSL.viewModelOf(
     crossinline constructor: (T1, T2) -> R,
     noinline options: DefinitionOptions<R>? = null,
 ): KoinDefinition<R> = viewModel { new(constructor) }.onOptions(options)
@@ -64,7 +62,7 @@ inline fun <reified R : ViewModel, reified T1, reified T2> Module.viewModelOf(
 /**
  * @see viewModelOf
  */
-inline fun <reified R : ViewModel, reified T1, reified T2, reified T3> Module.viewModelOf(
+inline fun <reified R : ViewModel, reified T1, reified T2, reified T3> ScopeDSL.viewModelOf(
     crossinline constructor: (T1, T2, T3) -> R,
     noinline options: DefinitionOptions<R>? = null,
 ): KoinDefinition<R> = viewModel { new(constructor) }.onOptions(options)
@@ -72,7 +70,7 @@ inline fun <reified R : ViewModel, reified T1, reified T2, reified T3> Module.vi
 /**
  * @see viewModelOf
  */
-inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T4> Module.viewModelOf(
+inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T4> ScopeDSL.viewModelOf(
     crossinline constructor: (T1, T2, T3, T4) -> R,
     noinline options: DefinitionOptions<R>? = null,
 ): KoinDefinition<R> = viewModel { new(constructor) }.onOptions(options)
@@ -80,7 +78,7 @@ inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T
 /**
  * @see viewModelOf
  */
-inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T4, reified T5> Module.viewModelOf(
+inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T4, reified T5> ScopeDSL.viewModelOf(
     crossinline constructor: (T1, T2, T3, T4, T5) -> R,
     noinline options: DefinitionOptions<R>? = null,
 ): KoinDefinition<R> = viewModel { new(constructor) }.onOptions(options)
@@ -88,7 +86,7 @@ inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T
 /**
  * @see viewModelOf
  */
-inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T4, reified T5, reified T6> Module.viewModelOf(
+inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T4, reified T5, reified T6> ScopeDSL.viewModelOf(
     crossinline constructor: (T1, T2, T3, T4, T5, T6) -> R,
     noinline options: DefinitionOptions<R>? = null,
 ): KoinDefinition<R> = viewModel { new(constructor) }.onOptions(options)
@@ -96,7 +94,7 @@ inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T
 /**
  * @see viewModelOf
  */
-inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T4, reified T5, reified T6, reified T7> Module.viewModelOf(
+inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T4, reified T5, reified T6, reified T7> ScopeDSL.viewModelOf(
     crossinline constructor: (T1, T2, T3, T4, T5, T6, T7) -> R,
     noinline options: DefinitionOptions<R>? = null,
 ): KoinDefinition<R> = viewModel { new(constructor) }.onOptions(options)
@@ -104,7 +102,7 @@ inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T
 /**
  * @see viewModelOf
  */
-inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T4, reified T5, reified T6, reified T7, reified T8> Module.viewModelOf(
+inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T4, reified T5, reified T6, reified T7, reified T8> ScopeDSL.viewModelOf(
     crossinline constructor: (T1, T2, T3, T4, T5, T6, T7, T8) -> R,
     noinline options: DefinitionOptions<R>? = null,
 ): KoinDefinition<R> = viewModel { new(constructor) }.onOptions(options)
@@ -112,7 +110,7 @@ inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T
 /**
  * @see viewModelOf
  */
-inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T4, reified T5, reified T6, reified T7, reified T8, reified T9> Module.viewModelOf(
+inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T4, reified T5, reified T6, reified T7, reified T8, reified T9> ScopeDSL.viewModelOf(
     crossinline constructor: (T1, T2, T3, T4, T5, T6, T7, T8, T9) -> R,
     noinline options: DefinitionOptions<R>? = null,
 ): KoinDefinition<R> = viewModel { new(constructor) }.onOptions(options)
@@ -120,7 +118,7 @@ inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T
 /**
  * @see viewModelOf
  */
-inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T4, reified T5, reified T6, reified T7, reified T8, reified T9, reified T10> Module.viewModelOf(
+inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T4, reified T5, reified T6, reified T7, reified T8, reified T9, reified T10> ScopeDSL.viewModelOf(
     crossinline constructor: (T1, T2, T3, T4, T5, T6, T7, T8, T9, T10) -> R,
     noinline options: DefinitionOptions<R>? = null,
 ): KoinDefinition<R> = viewModel { new(constructor) }.onOptions(options)
@@ -128,7 +126,7 @@ inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T
 /**
  * @see viewModelOf
  */
-inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T4, reified T5, reified T6, reified T7, reified T8, reified T9, reified T10, reified T11> Module.viewModelOf(
+inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T4, reified T5, reified T6, reified T7, reified T8, reified T9, reified T10, reified T11> ScopeDSL.viewModelOf(
     crossinline constructor: (T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11) -> R,
     noinline options: DefinitionOptions<R>? = null,
 ): KoinDefinition<R> = viewModel { new(constructor) }.onOptions(options)
@@ -136,7 +134,7 @@ inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T
 /**
  * @see viewModelOf
  */
-inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T4, reified T5, reified T6, reified T7, reified T8, reified T9, reified T10, reified T11, reified T12> Module.viewModelOf(
+inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T4, reified T5, reified T6, reified T7, reified T8, reified T9, reified T10, reified T11, reified T12> ScopeDSL.viewModelOf(
     crossinline constructor: (T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12) -> R,
     noinline options: DefinitionOptions<R>? = null,
 ): KoinDefinition<R> = viewModel { new(constructor) }.onOptions(options)
@@ -144,7 +142,7 @@ inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T
 /**
  * @see viewModelOf
  */
-inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T4, reified T5, reified T6, reified T7, reified T8, reified T9, reified T10, reified T11, reified T12, reified T13> Module.viewModelOf(
+inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T4, reified T5, reified T6, reified T7, reified T8, reified T9, reified T10, reified T11, reified T12, reified T13> ScopeDSL.viewModelOf(
     crossinline constructor: (T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13) -> R,
     noinline options: DefinitionOptions<R>? = null,
 ): KoinDefinition<R> = viewModel { new(constructor) }.onOptions(options)
@@ -152,7 +150,7 @@ inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T
 /**
  * @see viewModelOf
  */
-inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T4, reified T5, reified T6, reified T7, reified T8, reified T9, reified T10, reified T11, reified T12, reified T13, reified T14> Module.viewModelOf(
+inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T4, reified T5, reified T6, reified T7, reified T8, reified T9, reified T10, reified T11, reified T12, reified T13, reified T14> ScopeDSL.viewModelOf(
     crossinline constructor: (T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14) -> R,
     noinline options: DefinitionOptions<R>? = null,
 ): KoinDefinition<R> = viewModel { new(constructor) }.onOptions(options)
@@ -160,7 +158,7 @@ inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T
 /**
  * @see viewModelOf
  */
-inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T4, reified T5, reified T6, reified T7, reified T8, reified T9, reified T10, reified T11, reified T12, reified T13, reified T14, reified T15> Module.viewModelOf(
+inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T4, reified T5, reified T6, reified T7, reified T8, reified T9, reified T10, reified T11, reified T12, reified T13, reified T14, reified T15> ScopeDSL.viewModelOf(
     crossinline constructor: (T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14, T15) -> R,
     noinline options: DefinitionOptions<R>? = null,
 ): KoinDefinition<R> = viewModel { new(constructor) }.onOptions(options)
@@ -168,7 +166,7 @@ inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T
 /**
  * @see viewModelOf
  */
-inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T4, reified T5, reified T6, reified T7, reified T8, reified T9, reified T10, reified T11, reified T12, reified T13, reified T14, reified T15, reified T16> Module.viewModelOf(
+inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T4, reified T5, reified T6, reified T7, reified T8, reified T9, reified T10, reified T11, reified T12, reified T13, reified T14, reified T15, reified T16> ScopeDSL.viewModelOf(
     crossinline constructor: (T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14, T15, T16) -> R,
     noinline options: DefinitionOptions<R>? = null,
 ): KoinDefinition<R> = viewModel { new(constructor) }.onOptions(options)
@@ -176,7 +174,7 @@ inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T
 /**
  * @see viewModelOf
  */
-inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T4, reified T5, reified T6, reified T7, reified T8, reified T9, reified T10, reified T11, reified T12, reified T13, reified T14, reified T15, reified T16, reified T17> Module.viewModelOf(
+inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T4, reified T5, reified T6, reified T7, reified T8, reified T9, reified T10, reified T11, reified T12, reified T13, reified T14, reified T15, reified T16, reified T17> ScopeDSL.viewModelOf(
     crossinline constructor: (T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14, T15, T16, T17) -> R,
     noinline options: DefinitionOptions<R>? = null,
 ): KoinDefinition<R> = viewModel { new(constructor) }.onOptions(options)
@@ -184,7 +182,7 @@ inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T
 /**
  * @see viewModelOf
  */
-inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T4, reified T5, reified T6, reified T7, reified T8, reified T9, reified T10, reified T11, reified T12, reified T13, reified T14, reified T15, reified T16, reified T17, reified T18> Module.viewModelOf(
+inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T4, reified T5, reified T6, reified T7, reified T8, reified T9, reified T10, reified T11, reified T12, reified T13, reified T14, reified T15, reified T16, reified T17, reified T18> ScopeDSL.viewModelOf(
     crossinline constructor: (T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14, T15, T16, T17, T18) -> R,
     noinline options: DefinitionOptions<R>? = null,
 ): KoinDefinition<R> = viewModel { new(constructor) }.onOptions(options)
@@ -192,7 +190,7 @@ inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T
 /**
  * @see viewModelOf
  */
-inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T4, reified T5, reified T6, reified T7, reified T8, reified T9, reified T10, reified T11, reified T12, reified T13, reified T14, reified T15, reified T16, reified T17, reified T18, reified T19> Module.viewModelOf(
+inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T4, reified T5, reified T6, reified T7, reified T8, reified T9, reified T10, reified T11, reified T12, reified T13, reified T14, reified T15, reified T16, reified T17, reified T18, reified T19> ScopeDSL.viewModelOf(
     crossinline constructor: (T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14, T15, T16, T17, T18, T19) -> R,
     noinline options: DefinitionOptions<R>? = null,
 ): KoinDefinition<R> = viewModel { new(constructor) }.onOptions(options)
@@ -200,7 +198,7 @@ inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T
 /**
  * @see viewModelOf
  */
-inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T4, reified T5, reified T6, reified T7, reified T8, reified T9, reified T10, reified T11, reified T12, reified T13, reified T14, reified T15, reified T16, reified T17, reified T18, reified T19, reified T20> Module.viewModelOf(
+inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T4, reified T5, reified T6, reified T7, reified T8, reified T9, reified T10, reified T11, reified T12, reified T13, reified T14, reified T15, reified T16, reified T17, reified T18, reified T19, reified T20> ScopeDSL.viewModelOf(
     crossinline constructor: (T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14, T15, T16, T17, T18, T19, T20) -> R,
     noinline options: DefinitionOptions<R>? = null,
 ): KoinDefinition<R> = viewModel { new(constructor) }.onOptions(options)
@@ -208,7 +206,7 @@ inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T
 /**
  * @see viewModelOf
  */
-inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T4, reified T5, reified T6, reified T7, reified T8, reified T9, reified T10, reified T11, reified T12, reified T13, reified T14, reified T15, reified T16, reified T17, reified T18, reified T19, reified T20, reified T21> Module.viewModelOf(
+inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T4, reified T5, reified T6, reified T7, reified T8, reified T9, reified T10, reified T11, reified T12, reified T13, reified T14, reified T15, reified T16, reified T17, reified T18, reified T19, reified T20, reified T21> ScopeDSL.viewModelOf(
     crossinline constructor: (T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14, T15, T16, T17, T18, T19, T20, T21) -> R,
     noinline options: DefinitionOptions<R>? = null,
 ): KoinDefinition<R> = viewModel { new(constructor) }.onOptions(options)
@@ -216,7 +214,7 @@ inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T
 /**
  * @see viewModelOf
  */
-inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T4, reified T5, reified T6, reified T7, reified T8, reified T9, reified T10, reified T11, reified T12, reified T13, reified T14, reified T15, reified T16, reified T17, reified T18, reified T19, reified T20, reified T21, reified T22> Module.viewModelOf(
+inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T4, reified T5, reified T6, reified T7, reified T8, reified T9, reified T10, reified T11, reified T12, reified T13, reified T14, reified T15, reified T16, reified T17, reified T18, reified T19, reified T20, reified T21, reified T22> ScopeDSL.viewModelOf(
     crossinline constructor: (T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14, T15, T16, T17, T18, T19, T20, T21, T22) -> R,
     noinline options: DefinitionOptions<R>? = null,
 ): KoinDefinition<R> = viewModel { new(constructor) }.onOptions(options)
